requirements.txt
asgiref==3.7.2
Django==5.0.2
django-autoslug==1.9.9
django-cleanup==8.1.0
django-environ==0.11.2
gunicorn==21.2.0
packaging==23.2
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
